feat: add useCharacters hook
Custom hook using TanStack Query for fetching character data with pagination
